requirements.txt
bidict==0.23.1
blinker==1.9.0
bonboneka==0.2.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.3
Flask-SocketIO==5.6.1
greenlet==3.3.2
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
Werkzeug==3.1.7
wsproto==1.3.2
